Move useRequestStatus hook to a dedicated file
